lots of code for the same result ;-)
Pushing the code the provides the required and available lists of card groups into the set-specific packages. Instead of just pulling N groups at random from the full list, there is now a framework for pulling them from a curated list. That curated list doesn't yet exist, but it will once I add in set-specific config file handling.

I also added player count as a CL option to generate, and with the new rules module I leverage that to at very least provide counts for how many card groups for the given number of players (as per the rules book)

Next up, move some of the data that supports that logic (from the rule book) to set-specific config files. And just beyond that, some user config which provides blacklists (of which I have used heavily, as certain schemes are unplayable solo)