Change in plans
At first I was going to dish off to the decision engine at each card class choice, but instead I'll compile all the "valid" game configurations and then dish *that* off to the decision engine. It can then rule out more if it wants, then weight them (if appropriate) and make a choice.

available and config modules went away, they were (essentially) replaced by the rules config in the main package, coupled with the default values in the theme-specific packages.

main: configuration directory no longer is verified/created here - since rules uses it, rules can create it; added some verbosity settings to the CLI; --include-set option is now "multiple," meaning it can be given multiple times and game configurations will be assembled from all of the included sets; replaced all the code that just chose the "count" number of card groups by random, and replaced it with a loop that grows and contracts to whatever I define in the validate() method inside the GameConfiguration class

new stuff:

components: right now, just holds the GameConfiguration class - but may include others, just in case the enums end up not being enough metadata for validation to work

rules: this will be the home for all the functions needed by the validate() method

util: simple utility module - loading theme-specific packages from string names seems to be a need across the code, so that's why I created this and added that method