Implementing validation / rules; new rules format
I'm going with a different rules format for now. First section is merely the base counts by player count. These are expected to ONLY be in the base rules. Second section is blacklisted schemes (by player count) - this section can exist in both base and house rules, and will be combined (reconsider house rules always trump?). And the last section (of the base rules) is the scheme rules section. This section has a sub section for each scheme in the legendary set that needs to alter the rules. Sub sections below the scheme index are: masterminds, villains, henchmen, enforcers, and heroes_in_the_villain_deck. Under each of those, the possible categories are: diff, required, and exclusive. diff is just a positive or negative number that affect the base count for that card class (-1 villain, +1 hero in the villain deck, etc.). required provides a list of required items for that card group. This means that these items are required, but may be a subset of the total number of configured groups for this card class. For instance, were a henchman to be required (a value such as [1] as seen in the Buffy scheme with the index 3), there could be three henchmen configured in the game (like in a 4 or 5 player game), and one would need to be that henchman. Were two henchmen to be required, both would need to be in the three configured. "exclusive" is essentially the opposite. All of the configured card groups of that class would need to be a subset of what is set out in the rules. For instance, were one hero to be added to the villain deck (by setting diff to 1), and exclusive were to offer 4 possible heroes, the hero must be one of those. Were two heroes to be added to the villain deck (by setting diff to 2), and exclusive were set to the same 4 possible heroes, both heroes configured would need to be in the list offered to be a valid game config. This last section can also be defined in house rules, but the house rules *trump* the base rules - so be careful to set all the values necessary. In our example, if you wanted to add a 5th hero that was possible to add to the villain deck, you would need to define all five. But hopefully soon there will be a CLI sub command for adding, editing, and removing such things.

Note, there will be an exclusive section available to house rules where blacklisted combos can be defined. This would be something like "don't play these two heroes together," or "don't play this hero with this scheme," or "don't play hero X with scheme y and mastermind z."

components: added "always leads" (but haven't implemented it yet); added logic for moving a config through the process of filling all the "slots" scheme through hero in villain deck - this is done through the methods that accept new items to the config; implemented validation for scheme and masterminds

main: added "always leads" to CLI, passing it through, but not yet using it

rules: implemented "count" methods; implemented blacklisted scheme method; implemented both required and exclusive mastermind methods

NOTE: Big Trouble does NOT work with this code yet!