moved create_directory method; typos
So, I moved the create_directory method out of rules and into util. Seems like a trivial change, but it is the first change in a refactoring. Here's what's happening:

It's come time to eliminate the buffy and big_trouble packages. They were always temporary in order to force me to code towards multiple sets, not align myself with just one. They'll be replaced with configuration files that are read in upon startup to assemble a "set-specific" namespace inside which instances of the game model will live. Configs will be packaged with the code and written to a user-specific configuration directory at first execution. There will also be a restore sub command that can pull them out of the installation and overwrite any changes the user may have made. The default rules data files will work this way too. So, the next code to come will be some setup magic for packaging the data files, the data files themselves, some utility functions for move files around, the "restore" sub command, the startup command which invokes that, and two new modules: sets and model (described more below).

The big part of the refactor will be the sets and model modules. sets will be the home of the set-specific namespaces and the instances of the model classes within. It will also house the utility methods for loading the sets from disk. model will be where the class definitions of Scheme, Mastermind, Henchmen, etc. live.

Moving to these will seriously effect the rules and components modules. I get that. We'll see what happens. I may even merge the model config file with the rules config file. All depends on if it makes things easier or harder.

I know the big_trouble types module is on the way out, since I just went on and on about it, but there were typos I didn't want to carry forward in case I did a bit of copy/paste in the future.