First part of refactor
Created a model of the cards/card groups (at least the evil ones at this point) - so deleted the set-specific packages.

Added data files for that - more to be added there, like the json blurbs from the now-deleted rules files.

Deleted the components file and the test file - something like these objects will come back at some point, but they don't currently fit in with the load-from config at the start approach I am taking.

added the data files and as package_data to the setup file

added a new exception for things that go wrong during initialization (reading the configs from file)

overhauled main for this initialization-first approach.

moved a lot of constants and function to the util module